requirements.txt
# Telematics Standalone Testing Dependencies
requests>=2.30.0
python-dotenv>=1.0.0
